apps/api/pyproject.toml
[project]
name = "beleggingsapp-api"
version = "0.1.0"
description = "FastAPI backend voor de beleggingsapp"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "pyyaml>=6.0",
    "redis[asyncio]>=5.0.0",
    "structlog>=24.0.0",
    "yfinance>=0.2.40",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
